Add a parser for JSON style codecs
Implements a parser for codec definitions which slightly resemble
json and allow for complex data transformations and structured
objects.

A codec definition is in the format of:

codec my_codec {
    val1: u8,
    val2: little u8,
    val3 {
        child_val1: u8,
        child_val2: u16
    }
}

And also supports Java annotation-style attributes, which allows for
implementation-specific behavior based on the codec definition,
which is necessary for language interoperability.

Signed-off-by: Gary Tierney <gary.tierney@gmx.com>